requirements.txt
Django~=3.2
beautifulsoup4~=4.9.3
black~=21.5b2
bleach~=3.3.0
celery~=5.2.2
colorama~=0.4.4
django-extensions~=3.1.2
django-filter~=2.4.0
django-phonenumber-field~=5.0.0
django_nose~=1.4.7
djangorestframework~=3.12.4
fuzzywuzzy~=0.18.0
google-auth~=1.28.0
markdown~=3.3.4
phonenumberslite~=8.12.21
pinocchio~=0.4.2
psycopg2-binary~=2.8.6
python-Levenshtein~=0.12.2
python-dateutil~=2.8.1
requests~=2.25.1
uwsgi~=2.0.20
